move CorsHandler to common
